Moved low-level assign() helper to its own module.
