+++ lib/context.js
@@ -7,6 +7,7 @@ var util = require('util');
 
 var logging = require('../logging');
 
+var assign = require('../util/assign');
 var stack = require('../util/stack');
 
 var HasLevel = require('./HasLevel');
@@ -28,20 +29,6 @@ if(process.env.LOG_MSG_DATA)
         console.warn("Error parsing LOG_MSG_DATA from environment: %s", exc.stack || exc);
     } // end try
 } // end if
-
-function assign(target)
-{
-    var sources = Array.prototype.slice.call(arguments, 1);
-    for(var sourceIdx = 0; sourceIdx < sources.length; sourceIdx++)
-    {
-        var source = sources[sourceIdx];
-        for(var key in source)
-        {
-            target[key] = source[key];
-        } // end for
-    } // end for
-    return target;
-} // end assign
 
 // --------------------------------------------------------------------------------------------------------------------
